LMA: Adding HomeWork 8 - FluentObjectTest
