--- src/main/java/RedmineHW/RedmineAbstract.java
@@ -1,15 +1,0 @@
-package RedmineHW;
-
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.support.PageFactory;
-
-
-public abstract class RedmineAbstract {
-
-    public RedmineAbstract(WebDriver driver) {
-        PageFactory.initElements(driver, this);
-    }
-
-
-
-}
