+++ src/main/java/pageobject/SeznamResultPage.java
@@ -1,0 +1,20 @@
+package pageobject;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.How;
+
+public class SeznamResultPage extends AbstractPage {
+
+    @FindBy(how = How.XPATH, using = "//*[@data-dot='title']")
+    private WebElement firstResultLink;
+
+    public SeznamResultPage(WebDriver driver) {
+        super(driver);
+    }
+
+    public void clickOnResultLink() {
+        firstResultLink.click();
+    }
+}
