+++ src/main/java/redminehw/RedmineAbstract.java
@@ -1,0 +1,15 @@
+package redminehw;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.PageFactory;
+
+
+public abstract class RedmineAbstract {
+
+    public RedmineAbstract(WebDriver driver) {
+        PageFactory.initElements(driver, this);
+    }
+
+
+
+}
