+++ src/test/java/lma/fluenthwtest/FluentTest.java
@@ -1,0 +1,22 @@
+package lma.fluenthwtest;
+
+import fluenthw.FluentHomePage;
+import org.testng.annotations.Test;
+
+@Test
+public class FluentTest extends FluentTestAbstract {
+
+    @Test
+    public void FluentTest(){
+        getDriver().get("http://www.redmine.org/");
+
+        new FluentHomePage(getDriver())
+                .fluentLoginButton()
+                .fluentUsernameLogin("pepe5565")
+                .fluentUsernamePassword("PXPkehadBWX97Y9")
+                .fluentClickLoginButton()
+                .fluentHelpButton()
+                .fluentTimeReport()
+                .fluentTimeReportCheck();
+    }
+}
